add prop-types
prop-types will be used for type-checking for our props.
add initial draft of props to interface components.
re-organize pricing table components.
