+++ package.json
@@ -6,6 +6,7 @@
 	"dependencies": {
 		"formik": "^1.5.4",
 		"node-sass": "^4.11.0",
+		"prop-types": "^15.7.2",
 		"react": "^16.8.6",
 		"react-dom": "^16.8.6",
 		"react-ga": "^2.5.7",
